+++ app/cli/__init__.py
@@ -1,0 +1,1 @@
+"""Seiton Brain CLI-Paket (E16-3)."""